+++ lab6/zad4/Program.cs
@@ -12,12 +12,12 @@ namespace lab06
             Graph roads, paths, paths0, paths1, paths2;
             int[] cityCosts;
             string[] names;
-            List<List<int>> listOfResults0 = new List<List<int>>();
-            List<List<int>> listOfResults1 = new List<List<int>>();
-            List<List<int>> listOfResults2 = new List<List<int>>();
-            List<int> result0 = new List<int>();
-            List<int> result1 = new List<int>();
-            List<int> result2 = new List<int>();
+            List<List<double>> listOfResults0 = new List<List<double>>();
+            List<List<double>> listOfResults1 = new List<List<double>>();
+            List<List<double>> listOfResults2 = new List<List<double>>();
+            List<double> result0 = new List<double>();
+            List<double> result1 = new List<double>();
+            List<double> result2 = new List<double>();
             List<Graph> graphs = new List<Graph>();
             List<Graph> listOfPaths0 = new List<Graph>();
             List<Graph> listOfPaths1 = new List<Graph>();
@@ -26,7 +26,7 @@ namespace lab06
             List<int> bounds = new List<int>();
             List<int> expCosts = new List<int>();
             ulong cr;
-            ulong [,] cgg = new ulong[3,5];
+            ulong[,] cgg = new ulong[3, 5];
             cgg[0, 0] = 325;
             cgg[0, 1] = 211;
             cgg[0, 2] = 169;
@@ -53,7 +53,7 @@ namespace lab06
             listOfResults0.Add(result0);
             listOfResults1.Add(result1);
             listOfResults2.Add(result2);
-            
+
 
             Test1(out roads, out cityCosts, out result0, out result1, out result2, out paths0, out paths1, out paths2);
             graphs.Add(roads);
@@ -89,143 +89,143 @@ namespace lab06
 
             int n = 200;
             RandomGraphGenerator rgg = new RandomGraphGenerator(12345);
-            roads = rgg.UndirectedGraph(typeof(AdjacencyMatrixGraph),n,0.75,1,99);
+            roads = rgg.UndirectedGraph(typeof(AdjacencyMatrixGraph), n, 0.75, 1, 99);
             Random rnd = new Random(1234);
             cityCosts = new int[n];
-            for(int i=0; i<n;i++)
-                cityCosts[i]=rnd.Next(1,9);
+            for (int i = 0; i < n; i++)
+                cityCosts[i] = rnd.Next(1, 9);
             graphs.Add(roads);
             costs.Add(cityCosts);
 
             Pathfinder pf;
-            int minCost = 0;
-            int[] path;
+            double minCost = 0;
+            double[] path;
             bool pass;
 
             Console.WriteLine("\nCzesc 0\n");
-            for (int testno = 0; testno < graphs.Count-1; ++testno)
+            for (int testno = 0; testno < graphs.Count - 1; ++testno)
             {
-                    roads = graphs[testno];
-                    cityCosts = costs[testno];
-                    result0 = listOfResults0.ElementAt(testno);
+                roads = graphs[testno];
+                cityCosts = costs[testno];
+                result0 = listOfResults0.ElementAt(testno);
 
-                    names = new string[cityCosts.Length];
-                    for (int i = 0; i < cityCosts.Length; ++i)
-                        names[i] = String.Format("{0} ({1})", i, cityCosts[i]);
+                names = new string[cityCosts.Length];
+                for (int i = 0; i < cityCosts.Length; ++i)
+                    names[i] = String.Format("{0} ({1})", i, cityCosts[i]);
 
-                    pf = new Pathfinder(roads, cityCosts);
-                    minCost = 0;
-                    cr = Graph.Counter;
-                    path = pf.FindBestLocationWithoutCityCosts(out minCost, out paths);
-                    cr = Graph.Counter - cr;
+                pf = new Pathfinder(roads, cityCosts);
+                minCost = 0;
+                cr = Graph.Counter;
+                path = pf.FindBestLocationWithoutCityCosts(out minCost, out paths);
+                cr = Graph.Counter - cr;
 
-                    pass = path!=null && path.SequenceEqual(result0) && minCost == result0.Min();
-                    Console.WriteLine("Test {0} koszty  - {1}", testno, pass ? "Dobrze" : "Zle");
-                    if (paths != null)
-                    {
-                        pass = paths.IsEqual(listOfPaths0.ElementAt(testno));
-                    }
-                    else
-                    {
-                        pass = false;
-                    }
-                    Console.WriteLine("Test {0} sciezki - {1}", testno, pass ? "Dobrze" : "Zle");
-                    Console.WriteLine("Test {0} wydajnosc: {1} ( wzorcowa: {2} )", testno, cr, cgg[0, testno]);
+                pass = path != null && path.SequenceEqual(result0) && minCost == result0.Min();
+                Console.WriteLine("Test {0} koszty  - {1}", testno, pass ? "Dobrze" : "Zle");
+                if (paths != null)
+                {
+                    pass = paths.IsEqual(listOfPaths0.ElementAt(testno));
+                }
+                else
+                {
+                    pass = false;
+                }
+                Console.WriteLine("Test {0} sciezki - {1}", testno, pass ? "Dobrze" : "Zle");
+                Console.WriteLine("Test {0} wydajnosc: {1} ( wzorcowa: {2} )", testno, cr, cgg[0, testno]);
             }
 
-                    roads = graphs[graphs.Count-1];
-                    cityCosts = costs[graphs.Count-1];
-                    pf = new Pathfinder(roads, cityCosts);
-                    minCost = 0;
-                    cr = Graph.Counter;
-                    path = pf.FindBestLocationWithoutCityCosts(out minCost, out paths);
-                    cr = Graph.Counter - cr;
-                    Console.WriteLine("\nTest wydajnosci : {0} ( wzorcowo: {1} )", cr, cgg[0, 4]);
+            roads = graphs[graphs.Count - 1];
+            cityCosts = costs[graphs.Count - 1];
+            pf = new Pathfinder(roads, cityCosts);
+            minCost = 0;
+            cr = Graph.Counter;
+            path = pf.FindBestLocationWithoutCityCosts(out minCost, out paths);
+            cr = Graph.Counter - cr;
+            Console.WriteLine("\nTest wydajnosci : {0} ( wzorcowo: {1} )", cr, cgg[0, 4]);
 
             Console.WriteLine("\nCzesc 1\n");
-            for (int testno = 0; testno < graphs.Count-1; ++testno)
+            for (int testno = 0; testno < graphs.Count - 1; ++testno)
             {
-                    roads = graphs[testno];
-                    cityCosts = costs[testno];
-                    result1 = listOfResults1.ElementAt(testno);
+                roads = graphs[testno];
+                cityCosts = costs[testno];
+                result1 = listOfResults1.ElementAt(testno);
 
-                    names = new string[cityCosts.Length];
-                    for (int i = 0; i < cityCosts.Length; ++i)
-                        names[i] = String.Format("{0} ({1})", i, cityCosts[i]);
+                names = new string[cityCosts.Length];
+                for (int i = 0; i < cityCosts.Length; ++i)
+                    names[i] = String.Format("{0} ({1})", i, cityCosts[i]);
 
-                    pf = new Pathfinder(roads, cityCosts);
-                    minCost = 0;
-                    cr = Graph.Counter;
-                    path = pf.FindBestLocation(out minCost, out paths);
-                    cr = Graph.Counter - cr;
+                pf = new Pathfinder(roads, cityCosts);
+                minCost = 0;
+                cr = Graph.Counter;
+                path = pf.FindBestLocation(out minCost, out paths);
+                cr = Graph.Counter - cr;
 
-                    pass = path!=null && path.SequenceEqual(result1) && minCost == result1.Min();
-                    Console.WriteLine("Test {0} koszty  - {1}", testno, pass ? "Dobrze" : "Zle");
-                    if (paths != null)
-                    {
-                        pass = paths.IsEqual(listOfPaths1.ElementAt(testno));
-                    }
-                    else
-                    {
-                        pass = false;
-                    }
-                    Console.WriteLine("Test {0} sciezki - {1}", testno, pass ? "Dobrze" : "Zle");
-                    Console.WriteLine("Test {0} wydajnosc: {1} ( wzorcowa: {2} )", testno, cr, cgg[1, testno]);
+                pass = path != null && path.SequenceEqual(result1) && minCost == result1.Min();
+                Console.WriteLine("Test {0} koszty  - {1}", testno, pass ? "Dobrze" : "Zle");
+                if (paths != null)
+                {
+                    pass = paths.IsEqual(listOfPaths1.ElementAt(testno));
+                }
+                else
+                {
+                    pass = false;
+                }
+                Console.WriteLine("Test {0} sciezki - {1}", testno, pass ? "Dobrze" : "Zle");
+                Console.WriteLine("Test {0} wydajnosc: {1} ( wzorcowa: {2} )", testno, cr, cgg[1, testno]);
             }
 
-                    roads = graphs[graphs.Count-1];
-                    cityCosts = costs[graphs.Count-1];
-                    pf = new Pathfinder(roads, cityCosts);
-                    minCost = 0;
-                    cr = Graph.Counter;
-                    path = pf.FindBestLocation(out minCost, out paths);
-                    cr = Graph.Counter - cr;
-                    Console.WriteLine("\nTest wydajnosci : {0} ( wzorcowo: {1} )", cr, cgg[1, 4]);
+            roads = graphs[graphs.Count - 1];
+            cityCosts = costs[graphs.Count - 1];
+            pf = new Pathfinder(roads, cityCosts);
+            minCost = 0;
+            cr = Graph.Counter;
+            path = pf.FindBestLocation(out minCost, out paths);
+            cr = Graph.Counter - cr;
+            Console.WriteLine("\nTest wydajnosci : {0} ( wzorcowo: {1} )", cr, cgg[1, 4]);
 
             Console.WriteLine("\nCzesc 2\n");
-            for (int testno = 0; testno < graphs.Count-1; ++testno)
+            for (int testno = 0; testno < graphs.Count - 1; ++testno)
             {
-                    roads = graphs[testno];
-                    cityCosts = costs[testno];
-                    result2 = listOfResults2.ElementAt(testno);
+                roads = graphs[testno];
+                cityCosts = costs[testno];
+                result2 = listOfResults2.ElementAt(testno);
 
-                    names = new string[cityCosts.Length];
-                    for (int i = 0; i < cityCosts.Length; ++i)
-                        names[i] = String.Format("{0} ({1})", i, cityCosts[i]);
+                names = new string[cityCosts.Length];
+                for (int i = 0; i < cityCosts.Length; ++i)
+                    names[i] = String.Format("{0} ({1})", i, cityCosts[i]);
 
-                    minCost = 0;
-                    pf = new Pathfinder(roads, cityCosts);
-                    cr = Graph.Counter;
-                    path = pf.FindBestLocationSecondMetric(out minCost, out paths);
-                    cr = Graph.Counter - cr;
+                minCost = 0;
+                pf = new Pathfinder(roads, cityCosts);
+                cr = Graph.Counter;
+                path = pf.FindBestLocationSecondMetric(out minCost, out paths);
+                cr = Graph.Counter - cr;
 
-                    pass = path!=null && path.SequenceEqual(result2) && minCost == result2.Min();
-                    Console.WriteLine("Test {0} koszty  - {1}", testno, pass ? "Dobrze" : "Zle");
-                    if (paths != null)
-                    {
-                        pass = paths.IsEqual(listOfPaths2.ElementAt(testno));
-                    }
-                    else
-                    {
-                        pass = false;
-                    }
-                    Console.WriteLine("Test {0} sciezki - {1}", testno, pass ? "Dobrze" : "Zle");
-                    Console.WriteLine("Test {0} wydajnosc: {1} ( wzorcowa: {2} )", testno, cr, cgg[2, testno]);
+                pass = path != null && path.SequenceEqual(result2) && minCost == result2.Min();
+                Console.WriteLine("Test {0} koszty  - {1}", testno, pass ? "Dobrze" : "Zle");
+                if (paths != null)
+                {
+                    pass = paths.IsEqual(listOfPaths2.ElementAt(testno));
+                }
+                else
+                {
+                    pass = false;
+                }
+                Console.WriteLine("Test {0} sciezki - {1}", testno, pass ? "Dobrze" : "Zle");
+                Console.WriteLine("Test {0} wydajnosc: {1} ( wzorcowa: {2} )", testno, cr, cgg[2, testno]);
             }
 
-                    roads = graphs[graphs.Count-1];
-                    cityCosts = costs[graphs.Count-1];
-                    pf = new Pathfinder(roads, cityCosts);
-                    minCost = 0;
-                    cr = Graph.Counter;
-                    path = pf.FindBestLocationSecondMetric(out minCost, out paths);
-                    cr = Graph.Counter - cr;
-                    Console.WriteLine("\nTest wydajnosci : {0} ( wzorcowo: {1} )", cr, cgg[2, 4]);
+            roads = graphs[graphs.Count - 1];
+            cityCosts = costs[graphs.Count - 1];
+            pf = new Pathfinder(roads, cityCosts);
+            minCost = 0;
+            cr = Graph.Counter;
+            path = pf.FindBestLocationSecondMetric(out minCost, out paths);
+            cr = Graph.Counter - cr;
+            Console.WriteLine("\nTest wydajnosci : {0} ( wzorcowo: {1} )", cr, cgg[2, 4]);
 
             Console.WriteLine();
         }
 
-        static void Test0(out Graph roads, out int[] cityCosts, out List<int> result0, out List<int> result1, out List<int> result2, out Graph paths0 , out Graph paths1, out Graph paths2)
+        static void Test0(out Graph roads, out int[] cityCosts, out List<double> result0, out List<double> result1, out List<double> result2, out Graph paths0, out Graph paths1, out Graph paths2)
         {
             int n = 4;
 
@@ -235,38 +235,38 @@ namespace lab06
             cityCosts[1] = 1;
             cityCosts[2] = 100;
             cityCosts[3] = 2;
-            
+
             roads.AddEdge(0, 1, 4);
             roads.AddEdge(1, 3, 4);
             roads.AddEdge(0, 2, 1);
             roads.AddEdge(2, 3, 1);
 
-            result0 = new List<int>();
+            result0 = new List<double>();
             result0.Add(7);
             result0.Add(13);
             result0.Add(7);
             result0.Add(7);
 
-            result1 = new List<int>();
+            result1 = new List<double>();
             result1.Add(117);
             result1.Add(119);
             result1.Add(14);
             result1.Add(117);
 
-            result2 = new List<int>();
+            result2 = new List<double>();
             result2.Add(109);
             result2.Add(115);
             result2.Add(10);
             result2.Add(109);
 
-            paths0 = roads.IsolatedVerticesGraph(true,roads.VerticesCount);
-            paths0.AddEdge(1,0,4);
-            paths0.AddEdge(2,0,1);
+            paths0 = roads.IsolatedVerticesGraph(true, roads.VerticesCount);
+            paths0.AddEdge(1, 0, 4);
+            paths0.AddEdge(2, 0, 1);
             paths0.AddEdge(3, 2, 1);
 
             paths1 = roads.IsolatedVerticesGraph(true, roads.VerticesCount);
             paths1.AddEdge(0, 2, 1);
-            paths1.AddEdge(1,0,4);
+            paths1.AddEdge(1, 0, 4);
             paths1.AddEdge(3, 2, 1);
 
             paths2 = paths1;
@@ -276,7 +276,7 @@ namespace lab06
 
 
 
-        static void Test1(out Graph roads, out int[] cityCosts, out List<int> result0, out List<int> result1, out List<int> result2, out Graph paths0, out Graph paths1, out Graph paths2)
+        static void Test1(out Graph roads, out int[] cityCosts, out List<double> result0, out List<double> result1, out List<double> result2, out Graph paths0, out Graph paths1, out Graph paths2)
         {
             int n = 5;
 
@@ -288,8 +288,8 @@ namespace lab06
             cityCosts[3] = 3;
             cityCosts[4] = 4;
 
-            
-            result1 = new List<int>();
+
+            result1 = new List<double>();
             result1.Add(40000);
             result1.Add(40000);
             result1.Add(40000);
@@ -309,7 +309,7 @@ namespace lab06
             paths2 = paths0;
         }
 
-        static void Test2(out Graph roads, out int[] cityCosts, out List<int> result0, out List<int> result1, out List<int> result2, out Graph paths0, out Graph paths1, out Graph paths2)
+        static void Test2(out Graph roads, out int[] cityCosts, out List<double> result0, out List<double> result1, out List<double> result2, out Graph paths0, out Graph paths1, out Graph paths2)
         {
             int n = 3;
 
@@ -323,22 +323,22 @@ namespace lab06
             roads.AddEdge(1, 2, 5);
             roads.AddEdge(0, 2, 5);
 
-            result0 = new List<int>();
+            result0 = new List<double>();
             result0.Add(15);
             result0.Add(15);
             result0.Add(10);
 
-            result1 = new List<int>();
+            result1 = new List<double>();
             result1.Add(27);
             result1.Add(22);
             result1.Add(25);
 
-            result2 = new List<int>();
+            result2 = new List<double>();
             result2.Add(22);
             result2.Add(12);
             result2.Add(25);
 
-            paths0 = roads.IsolatedVerticesGraph(true,roads.VerticesCount);
+            paths0 = roads.IsolatedVerticesGraph(true, roads.VerticesCount);
             paths0.AddEdge(0, 2, 5);
             paths0.AddEdge(1, 2, 5);
 
@@ -347,13 +347,13 @@ namespace lab06
             paths1.AddEdge(2, 1, 5);
 
             paths2 = roads.IsolatedVerticesGraph(true, roads.VerticesCount);
-            paths2.AddEdge(0,1, 100);
+            paths2.AddEdge(0, 1, 100);
             paths2.AddEdge(2, 1, 5);
 
         }
 
 
-        static void Test3(out Graph roads, out int[] cityCosts, out List<int> result0, out List<int> result1, out List<int> result2, out Graph paths0, out Graph paths1, out Graph paths2)
+        static void Test3(out Graph roads, out int[] cityCosts, out List<double> result0, out List<double> result1, out List<double> result2, out Graph paths0, out Graph paths1, out Graph paths2)
         {
             int n = 5;
             roads = new AdjacencyMatrixGraph(false, n);
@@ -370,21 +370,21 @@ namespace lab06
             roads.AddEdge(3, 4, 5);
             roads.AddEdge(4, 3, 5);
 
-            result0 = new List<int>();
+            result0 = new List<double>();
             result0.Add(20015);
             result0.Add(20015);
             result0.Add(20010);
             result0.Add(30005);
             result0.Add(30005);
 
-            result1 = new List<int>();
+            result1 = new List<double>();
             result1.Add(20027);
             result1.Add(20022);
             result1.Add(20025);
             result1.Add(30012);
             result1.Add(30012);
 
-            result2 = new List<int>();
+            result2 = new List<double>();
             result2.Add(20022);
             result2.Add(20012);
             result2.Add(20025);
